requirements.txt
pandas==1.5.3
numpy==1.24.3
streamlit
scikit-learn==1.2.2
joblib==1.2.0
Pillow
